requirements.txt
flask==3.1.0
gunicorn==23.0.0
libsql
python-dotenv==1.1.0
bcrypt==4.3.0
openpyxl==3.1.5
